test(windows): fix TestQueryUpdateFromIDEditor
Our Windows CI pipeline is broken, this fixes the following error:
```
=== FAIL: integration TestQueryUpdateFromIDEditor (1.37s)
panic: unsupported [recovered]
        panic: unsupported
```
