+++ integration/lql_update_test.go
@@ -23,9 +23,7 @@ import (
 	"fmt"
 	"os"
 	"testing"
-	"time"
 
-	"github.com/Netflix/go-expect"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -101,52 +99,4 @@ func TestQueryUpdateFromIDNotFound(t *testing.T) {
 	assert.Contains(t, stderr.String(), "/api/v2/Queries/ID_NOT_FOUND")
 	assert.Contains(t, stderr.String(), "Query id ID_NOT_FOUND not found")
 	assert.Equal(t, 1, exitcode, "EXITCODE is not the expected one")
-}
-
-func TestQueryUpdateFromIDEditor(t *testing.T) {
-	// get temp file
-	file, err := createTemporaryFile(
-		"TestQueryUpdateFromIDEditor",
-		fmt.Sprintf(queryJSONTemplate, evaluatorID, queryID, queryUpdateText),
-	)
-	if err != nil {
-		t.FailNow()
-	}
-	defer os.Remove(file.Name())
-
-	dir := createTOMLConfigFromCIvars()
-	defer os.RemoveAll(dir)
-
-	// setup
-	LaceworkCLIWithHome(dir, "query", "create", "-u", queryURL)
-	// teardown
-	defer LaceworkCLIWithHome(dir, "query", "delete", queryID)
-
-	_ = runFakeTerminalTestFromDir(t, dir,
-		func(c *expect.Console) {
-			expectString(t, c, "Update query")
-			c.SendLine("")
-			time.Sleep(time.Millisecond)
-			// Move to line number 4 and add comment "--- Updated from CLI Editor"
-			c.Send("4Go--- Updated from CLI Editor\x1b")
-			c.SendLine(":wq!") // save and close
-			time.Sleep(time.Millisecond)
-			expectString(t, c,
-				fmt.Sprintf("The query %s was updated.", queryID))
-		},
-		"query", "update", queryID,
-	)
-
-	t.Run("verify query editions", func(t *testing.T) {
-		stdout, stderr, exitcode := LaceworkCLIWithHome(dir, "query", "show", queryID)
-		assert.Empty(t,
-			stderr.String(),
-			"STDERR should be empty")
-		assert.Contains(t,
-			stdout.String(),
-			"--- Updated from CLI Editor",
-			"the query was not editted correctly")
-		assert.Equal(t, 0, exitcode,
-			"EXITCODE is not the expected one")
-	})
 }
